% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_scatter.R
\name{add_contour_to_pointcloud}
\alias{add_contour_to_pointcloud}
\title{Add contour line to a xy-plot that envelopes \code{alpha * 100}-percent
of the xy-points}
\usage{
add_contour_to_pointcloud(
  x,
  y,
  alpha = 0.95,
  col = "red",
  fill = FALSE,
  na.rm = FALSE,
  ...
)
}
\description{
Add contour line to a xy-plot that envelopes \code{alpha * 100}-percent
of the xy-points
}
\references{
Modified from code by Pascal Haenggi (2008)
  https://stat.ethz.ch/pipermail/r-help/2008-June/166079.html
}
