% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_scatter.R
\name{add_isoline_to_pointcloud}
\alias{add_isoline_to_pointcloud}
\title{Add isoline(s) to a \var{xy}-plot}
\usage{
add_isoline_to_pointcloud(
  x,
  y,
  z,
  probs = NULL,
  levels = NULL,
  col = "red",
  fill = FALSE,
  na.rm = FALSE,
  add_legend = FALSE,
  legend_xy = NULL,
  ...
)
}
\description{
The isolines can either represent the \code{probs * 100}-percent
quantile or represent fixed \code{levels} of the z-variable values
}
