% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_maps.R
\name{add_panel_map}
\alias{add_panel_map}
\title{Create a raster and plot it to the current device}
\usage{
add_panel_map(x, meta, subset, ...)
}
\description{
Create a raster and plot it to the current device
}
