% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ensemble.R
\name{calc_cellwise_ensemble}
\alias{calc_cellwise_ensemble}
\title{Calculate cell-wise ensembles across \code{reqCSs}}
\usage{
calc_cellwise_ensemble(
  data,
  subset = NULL,
  funs = c("mean", "min", "max"),
  ...,
  variables = NULL,
  sc_historical = NULL,
  add_historical = FALSE,
  dim_ens = 3L,
  req_scens = NULL,
  req_dtimes = NULL,
  req_downs = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{An \code{array} with four dimensions -- as generated by
\code{\link{extract_dbOut_to_array}}.}

\item{subset}{A logical vector used to subset gridcells from \code{data}
for calculation of ensemble.}

\item{funs}{A vector of character strings or of functions. The functions of
the form \code{f(x, ...)} which will be used to calculate the ensemble
values across \var{GCMs}.}

\item{...}{Optional arguments passed to every \code{funs} function, e.g.,
\code{na.rm = TRUE}.}

\item{variables}{A vector of character strings. A subset of names of the
second dimension of \code{data} or \code{NULL} which indicates to use
the full set of the second dimension of \code{data}.}

\item{dim_ens}{An integer value. The dimension of \code{data} over which
the ensembles are calculated. This has to be \code{3}.}

\item{req_scens}{A vector of character strings. A subset of
(climate) scenarios from \code{SFSW2_prj_meta[["sim_scens"]][["reqCSs"]]}
for which ensembles are calculated across participating models.}
}
\description{
Calculate cell-wise ensembles across \code{reqCSs}
}
