% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_mathematical.R
\name{calc_NT1}
\alias{calc_NT1}
\title{Univariate novelty/extrapolation NT1}
\usage{
calc_NT1(refdat, prodat)
}
\arguments{
\item{refdat}{A numerical matrix. The reference values of variables organized
in columns.}

\item{prodat}{A numerical matrix. The projected values. The columns must
match those of \code{refdat}.}
}
\description{
"NT1 ranges from infinite negative values to zero where zero indicates no
extrapolation beyond the univariate coverage of reference data"
(Mesgaran et al. 2014).
}
\section{Notes}{
 The code is adapted from Bell & Schlaepfer 2015 (available
  at \url{https://github.com/bellland/SDM.Virtual.Species_Bell.Schlaepfer})
  which was based on a comment by Matthew Bayly made at
  \url{https://pvanb.wordpress.com/2014/05/13/a-new-method-and-tool-exdet-to-evaluate-novelty-environmental-conditions/}.
}

\references{
Mesgaran, M. B., R. D. Cousens, B. L. Webber, and J. Franklin.
  2014. Here be dragons: a tool for quantifying novelty due to covariate
  range and correlation change when projecting species distribution models.
  Diversity and Distributions 20:1147-1159.
}
