% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ensemble.R
\name{calc_regionwise_ensemble}
\alias{calc_regionwise_ensemble}
\title{Assign to cells the \var{inverse-ecdf} quantiles from region-wide ranked GCMs
ensembles across \code{reqCSs}}
\usage{
calc_regionwise_ensemble(
  SFSW2_prj_meta,
  data,
  subset = NULL,
  region,
  area = NULL,
  fcentral = c("median", "mean"),
  funs = c("min", "median", "max"),
  probs = c(0, 0.5, 1),
  ...,
  ties.method = "first",
  variables = NULL,
  add_historic = TRUE,
  reqCSs = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data.frame with three dimensions -- as generated by
\code{\link{extract_dbOut_to_array}}.}

\item{subset}{A logical vector used to subset gridcells from \code{data}
for calculation of ensemble.}

\item{region}{A character or numeric vector of a length equal to the first
dimension of \code{data}, i.e., cells.}

\item{area}{A numeric vector. Its length corresponds to the first dimension
of \code{data} and represents the cell areas (e.g., that may vary by
latitude). If not \code{NULL} or not all equal, then a \code{fcentral}
value "mean" uses \code{\link[stats]{weighted.mean}} instead of
\code{\link[base]{mean}} and a value of "median" uses
\code{\link[Hmisc]{wtd.quantile}} instead of \code{\link[stats]{median}}
where \code{area} are used as non-random "reliability" weights.}

\item{fcentral}{A character string naming the function to calculate a central
tendency across the spatial extent for each \var{GCMs};
one of "mean" or "median". The function is used to calculate the
region-wide (weighted) statistic on which the GCMs are ranked. See
\code{area}.}

\item{funs}{A vector of character strings or of functions. The functions of
the form \code{f(x, ...)} which will be used to calculate the ensemble
values across \var{GCMs}. Note: \var{"min"}, \var{"median"}, \var{"max"}
will be replaced by their quantile equivalents,
i.e., \code{probs = 0, 0.5, or 1}  respectively.}

\item{probs}{A numeric vector of probabilities with values in \code{[0,1]}
to determine which quantiles from among the \code{fcentral} values
across \var{GCMs} are returned for each cell.}

\item{...}{Optional arguments passed to \code{\link[stats]{quantile}}, e.g.,
\code{na.rm = TRUE}.}

\item{ties.method}{A character string. Specifies how ties are treated. See
\code{link[base]{rank}}. The method must produce unique ranks if
\code{funs} or \code{probs} are not empty.}

\item{variables}{A vector of character strings. A subset of names of the
second dimension of \code{data} or \code{NULL} which indicates to use the
full set of the second dimension of \code{data}.}

\item{reqCSs}{A vector of character strings. A subset of (climate) scenarios
from \code{SFSW2_prj_meta[["sim_scens"]][["reqCSs"]]} for which ensembles
are calculated.}

\item{verbose}{A logical value.}
}
\description{
For each \code{region}, the function \code{\link{calc_extentwise_ensemble}}
is called on the set of cells belonging to that region.
}
\seealso{
\code{\link{calc_extentwise_ensemble}}
}
