% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_change.R
\name{calc_response_change_from_reference}
\alias{calc_response_change_from_reference}
\title{Calculate absolute or relative change in response variables}
\usage{
calc_response_change_from_reference(
  data,
  ref_condition,
  sc_hist = NULL,
  method = c("absolute", "relative", "direction", "rdirection"),
  ...
)
}
\arguments{
\item{data}{A four-dimensional numerical \code{array} with dimensions
\code{runsN_sites}, \code{variables}, \code{scenarios},
and \code{experiments}, e.g., the returned object from
\code{\link{extract_dbOut_to_array}}.}

\item{ref_condition}{A vector of character strings. Each element is used
as reference to calculate a separate set of change values.}

\item{sc_hist}{A character string. The name of the historical scenario that
is the same for each other (future) projection; or \code{NULL}.}

\item{method}{A character string. See description.}

\item{...}{Additional arguments, e.g., \describe{
\item{tol}{two-sided tolerance in relative units to identify
      "no change" for \code{method == "direction"} and
      \code{method == "rdirection"}}
\item{subset}{a logical vector of length \code{runsN_sites} used by
      \code{method == "rdirection"} to identify "no change"}
}}
}
\value{
A list of four-dimensional numerical \code{array} with dimensions
  \code{runsN_sites}, \code{variables}, \code{scenarios},
  and \code{experiments}; each element is named as \var{delta_ref_met}
  where \var{ref} is one of \code{ref_condition} and \var{met} is the
  first three letters of \code{method} and the value of \code{tol} if
  provided.
}
\description{
Change in response variables between a climate reference
condition (or scenario family) and each climate condition is calculated as
\itemize{
  \item \code{x[condition k] - x[reference]}
         if \code{method} is \code{"absolute"}
  \item \code{(x[condition k] - x[reference]) / x[reference]}
         if \code{method} is \code{"relative"}
  \item the result of \code{\link{compare_direction_2deltas}}
         if \code{method} is \code{"direction"}
  \item the result of \code{(x[condition k] > tol, x[condition k] < tol)}
         if \code{method} is relative direction, i.e., \code{"rdirection"}
}
}
\details{
If \code{reference} is the same for
each scenario, e.g., "current", then the same value is used for each
\code{condition k}; if \code{reference} is a scenario family, e.g., "RCP45",
then the appropriate \code{x} values per GCM are used.
}
