% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ensemble.R
\name{calculate_ensembles}
\alias{calculate_ensembles}
\title{Ensemble from ranked GCM at the scale of interest (Madsen et al. 2017)}
\usage{
calculate_ensembles(
  meta,
  data,
  data_names = names(data),
  subset,
  cell_area_km2,
  id_region,
  sc_historical,
  req_Downs,
  req_dTime,
  ens_wises = c("EnsCW", "EnsRW", "EnsGW"),
  ens_funs = c("min", "mean", "median", "max", "span", "agreement", "majority"),
  path,
  ftag
)
}
\description{
Ensemble from ranked GCM at the scale of interest (Madsen et al. 2017)
}
\references{
Madsen, M. S., P. L. Langen, F. Boberg, and J. H. Christensen. 2017.
  Inflated Uncertainty in Multimodel-Based Regional Climate Projections.
  Geophysical Research Letters 44:11606–11613.
}
