% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{calculate_nominal_resolution}
\alias{calculate_nominal_resolution}
\title{Calculate "nominal resolution" of grid}
\usage{
calculate_nominal_resolution(grid, sites, cell_areas_km2)
}
\description{
Calculate "nominal resolution" of grid
}
\references{
CMIP6 Global Attributes, DRS, Filenames, Directory Structure, and CV’s
  10 September 2018 (v6.2.7)
  Appendix 2: Algorithms for Defining the "nominal_resolution" Attribute
  https://docs.google.com/document/d/1h0r8RZr_f3-8egBMMh7aqLwy3snpD6_MrDz1q8n5XUk/edit#bookmark=id.ibeh7ad2gpdi
}
