% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_change.R
\name{compare_direction_1rel2ref}
\alias{compare_direction_1rel2ref}
\title{Compare directions of change}
\usage{
compare_direction_1rel2ref(
  dx,
  dy,
  vars = NULL,
  cl = NULL,
  tol = NULL,
  subset = NULL
)
}
\arguments{
\item{dx}{Reference data}
}
\description{
Compare directions of change
}
