% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{create_empty_netCDF_file}
\alias{create_empty_netCDF_file}
\title{create_empty_netCDF_file
Create a structured, empty netcdf file}
\usage{
create_empty_netCDF_file(
  data,
  has_T_timeAxis = FALSE,
  has_Z_verticalAxis = FALSE,
  time_bounds,
  vert_bounds,
  var_attributes,
  time_attributes,
  vertical_attributes,
  global_attributes,
  isGridded = TRUE,
  grid = NULL,
  locations,
  crs,
  file,
  force_v4 = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{data}{A numeric array.}

\item{has_T_timeAxis}{A logical value. Indicates that the netCDF created will
have a third, time dimension.}

\item{has_Z_verticalAxis}{A logical value. Indicates that the netCDF created will
have a veritical (e.g. soil profile depths) dimension. If \var{has_T_timeAxis}
is set to \code{FALSE} then the Z axis will be the third dimension.
If \var{has_T_timeAxis} is set to \code{TRUE} then the Z axis will be the
fourth dimension.}

\item{time_bounds}{A numeric string that continuously lists the lower and upper
bounds of each time dimension value. In the absence of a time dimension,
this argument can be used to define the calculation period of the variable(s).}

\item{vert_bounds}{A numeric string that continuously lists the lower and upper
bounds of each vertical dimension value.}

\item{var_attributes}{A list of named character strings defining the variables
of the netCDF.}

\item{time_attributes}{A list of named character strings defining the time
dimension of the netCDF.}

\item{vertical_attributes}{A list of named character strings defining the vertical
dimension of the netCDF.}

\item{global_attributes}{A list of named character strings defining the
global attributes of the netCDF.
then the global attributes will be added to the netCDF file.}

\item{grid}{filename (character). File containing the grid information
(i.e. resolution, extent, locations) of the data. Supported file types are
the 'native' raster package format and those that can be read via rgdal.}

\item{file}{A character string. The file path of the netCDF file to be
created.}

\item{force_v4}{A logical value. Force version 4 of netCDF.}

\item{overwrite}{A logical value. If \code{TRUE}, file will be overwritten
if it already exists.}

\item{isGridded.}{A logical value. Represents whether the location data is on
a regular grid or not.}

\item{locations.}{A SpatialPoints object or a matrix or data.frame with two
columns containing long and lat values. Data must be organized by long, lat.}

\item{crs.}{character or object of class 'CRS'.}
}
\value{
This function is used for the side-effect of creating a file.
}
\description{
This function creates a strucutured netCDF with metadata, but
 without data. The user must specific spatial information, dimensionality, and
 metadata of the netCDF. The dimensions of the netCDF always includes at least two dimensions,
 Latitude and Longitude, with information about these dimensions gathered from
 spatial data (i.e. \var{locations} or \var{grid}).
 Additional dimensionalilty information is gathered through the use of the vars
 \var{has_T_timeAxis} & \var{has_Z_verticalAxis}. If both flags are set to
 \code{FALSE} then the netCDF is two dimensions with one or more variables.
 The number of variables is set in the \var{var_attributes} argument. If either
 \var{has_T_timeAxis} \emph{or} \var{has_Z_verticalAxis} is set to \code{TRUE} then a
 third dimension is added to the netCDF. If \emph{both} \var{has_T_timeAxis}
 and \var{has_Z_verticalAxis} are set to TRUE, then the netCDF will have 4
 dimensions, with time as the third, and depth as the fourth. Information about
 these dimensions is set in the respective \var{bounds} and \var{attributes}.
 \var{has_T_timeAxis} and/or \var{has_Z_verticalAxis} is set to \code{TRUE}
 the netCDF can \emph{only} have one variable.
}
\examples{

}
