% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{create_netCDF_from_raster_with_time}
\alias{create_netCDF_from_raster_with_time}
\title{Write raster where a third dimension represents time to a netCDF file}
\usage{
create_netCDF_from_raster_with_time(x, file, ...)
}
\arguments{
\item{x}{A Raster* object}
}
\description{
Write raster where a third dimension represents time to a netCDF file
}
