% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{create_netCDF_from_raster_with_variables}
\alias{create_netCDF_from_raster_with_variables}
\alias{create_netCDF_from_array_with_variables}
\title{Convert raster where variables are organized in the third dimension to a
\var{netCDF} file}
\usage{
create_netCDF_from_raster_with_variables(
  x,
  time_bounds,
  var_attributes,
  global_attributes,
  file,
  force_v4 = TRUE,
  overwrite = FALSE
)

create_netCDF_from_array_with_variables(
  x,
  locations,
  grid,
  time_bounds,
  var_attributes,
  global_attributes,
  file,
  force_v4 = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{x}{A Raster* object.}

\item{var_attributes}{A list of named character strings each of length
equal to the number of variables, i.e., the third dimension of \code{x}.
If not missing, then the variable attributes will be added to the
\var{netCDF} file. If no variable names are provided, then the names of
\code{x} are used; otherwise, these are added as \code{original_name}.}

\item{global_attributes}{A list of named character strings. If not missing,
then the global attributes will be added to the \var{netCDF} file.}

\item{file}{A character string. The file path of the \var{netCDF} file to be
created.}

\item{overwrite}{A logical value. If \code{TRUE}, file will be overwritten
  if it exists.

 @return This function is used for the side-effect of creating a file.}
}
\description{
Convert raster where variables are organized in the third dimension to a
\var{netCDF} file
}
\section{Functions}{
\itemize{
\item \code{create_netCDF_from_array_with_variables}: Convert array where
variables are organized in the third dimension to a \var{netCDF} file
}}

