% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_functions_miscellaneous.R
\name{define_delta_colorramp}
\alias{define_delta_colorramp}
\title{Define a colorramp based one set of colors for the negative range and
another set of colors for the positive range}
\usage{
define_delta_colorramp(
  zlim,
  zlim_trim = NULL,
  Nc = 255,
  use_rev_cols = FALSE,
  add_extremes_to_legend = FALSE,
  col_zero = "gray",
  cols_blues = c("royalblue", "cornflowerblue", "lightblue"),
  cols_reds = c("lightsalmon", "mediumorchid1", "purple", "purple4")
)
}
\description{
Define a colorramp based one set of colors for the negative range and
another set of colors for the positive range
}
