% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_rSFSW2_output.R
\name{extract_dbOut_to_array}
\alias{extract_dbOut_to_array}
\title{Extract data from \var{dbOutput}}
\usage{
extract_dbOut_to_array(
  meta,
  fname_dbOut = NULL,
  file = NULL,
  MeanOrSD = "Mean",
  runsN_sites = NULL,
  variables = NULL,
  scenarios = NULL,
  experiments = NULL,
  verbose = FALSE
)
}
\arguments{
\item{file}{A character string. The file path to the \var{rds}-file(s)
containing the memoised data. If \code{NULL}, then data is not written/read
from file, but extracted from \var{dbOutput}.}
}
\value{
A four-dimensional numerical \code{array} with dimensions
  \code{runsN_sites}, \code{variables}, \code{scenarios}, and
  \code{experiments}.
}
\description{
This function is internally "memoised" per variable, i.e., it loads an
\code{array} from \code{file} if previously extracted from
\var{dbOutput}; otherwise, it extracts first data (and then saves to a file
disk).
}
\seealso{
\code{\link{get.SeveralOverallVariables_Scenario}},
  \code{\link{saveRDS}}
}
