% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_change.R
\name{factor_6directions}
\alias{factor_6directions}
\title{Convert data to factor according to
\code{define_direction_2deltas()[, "name"]}}
\usage{
factor_6directions(data, cl = NULL)
}
\description{
Convert data to factor according to
\code{define_direction_2deltas()[, "name"]}
}
