% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_raster.R
\name{get_raster_datatype}
\alias{get_raster_datatype}
\title{Convert \code{\link{typeof}} to \code{\link[raster]{dataType}} types}
\usage{
get_raster_datatype(data)
}
\description{
Convert \code{\link{typeof}} to \code{\link[raster]{dataType}} types
}
\references{
Relevant code adapted from \code{`raster:::dataType<-`}
}
