% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geographical.R
\name{get_UTM_Zone}
\alias{get_UTM_Zone}
\title{Calculate the UTM zone based on geographic location}
\usage{
get_UTM_Zone(longitude, latitude)
}
\description{
Calculate the UTM zone based on geographic location
}
\references{
Convert Latitude/Longitude to UTM [https://www.wavemetrics.com/code-snippet/convert-latitudelongitude-utm] (attributed to Chuck Gantz).
}
