% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_rSFSW2_output.R
\name{load_rSFSW2_data_for_analysis}
\alias{load_rSFSW2_data_for_analysis}
\title{Prepare and load \var{SOILWAT2} simulation output as an array
from a \var{sqlite3} database produced by \pkg{rSFSW2}.}
\usage{
load_rSFSW2_data_for_analysis(
  meta,
  path,
  path_tmp,
  fname_dbOuts = NULL,
  variables,
  MeanOrSD = c("Mean", "SD"),
  sim_scenarios,
  sc_historical,
  experiments = NULL,
  subprojects = NULL,
  subset = NULL,
  write_to_netcdf = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{experiments}{A vector of character strings. Values of the
\var{'Experimental_Label'} field. Use if experimental
treatments are organized in one \var{dbOutput} database.}

\item{subprojects}{A vector of character strings. Same as \code{experiments},
but use if experimental treatments are organized in different
\var{dbOutput} databases. If used, then \code{fname_dbOuts} is a vector
with values for each \code{subprojects}. Note: \code{experiments} and
\code{subprojects} are mutually exclusive.}
}
\value{
The same structure as returned from
  \code{\link{extract_dbOut_to_array}}, i.e., a four-dimensional numerical
  \code{array} with dimensions \code{runsN_sites}, \code{variables},
  \code{scenarios}, and \code{experiments}.
}
\description{
Prepare and load \var{SOILWAT2} simulation output as an array
from a \var{sqlite3} database produced by \pkg{rSFSW2}.
}
\section{Notes}{
 \code{write_to_netcdf} is no longer supported.
}

\seealso{
\code{\link{extract_dbOut_to_array}} for extraction from
  \var{dbOut} database.
}
