% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_maps.R
\name{map_from_variable}
\alias{map_from_variable}
\title{Create a raster from values and plot it to the current device}
\usage{
map_from_variable(x, meta, subset, ..., beautify = TRUE)
}
\arguments{
\item{x}{A vector of values to plot as a map, see details.}

\item{meta}{The metadata list of a \pkg{rSFSW2} simulation experiment.}

\item{subset}{A vector logical vector to subset \code{x} or missing.}

\item{...}{Additional parameters most of which are passed to the plotting
function(s) \code{\link[raster]{image}} or
\code{\link[graphics]{rasterImage}};
exceptions are explained in the details.}

\item{beautify}{Add country borders and the equator to the plot.}
}
\value{
The raster object generated from \code{x} by a call to
  \code{\link{create_raster_from_variables}} is returned invisibly.
  The function is called for its side effect of plotting to the
  current device.
}
\description{
Create a raster from values and plot it to the current device
}
\details{
\code{x} can be specified in two ways:
 codes: \itemize{
   \item If \code{x} is a numerical vector, then the values are interpreted
     as \var{z} values and plotted by the function
     \code{\link[raster]{image}} which calls \code{\link[graphics]{image}}.
   \item If \code{x} is a vector of hexadecimal color codes, then the
     values are plotted by the function \code{\link[graphics]{rasterImage}},
     similar to how \code{\link[raster]{plotRGB}} plots multi-color rasters.
 }

The following arguments in \code{...} are treated separately:
\itemize{
  \item If \code{ylim} is available but not \code{asp}, then an aspect
    ratio is calculated that corrects for longitude/latitude.
  \item If \code{useNA} is \code{TRUE} and \code{valNA} is available, then
    \code{NA}s are replaced by \code{valNA}.
  \item If \code{trim} is \code{TRUE}, then the resulting raster object
    is trimmed using \code{\link[raster]{trim}}.
  \item If \code{fextend} is a finite number, then the extent of the raster
    object (after trimming) is expanded by that factor.
}
}
