% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{populate_netcdf_from_array}
\alias{populate_netcdf_from_array}
\title{populate_netcdf_from_array
Populate an empty netcdf with values}
\usage{
populate_netcdf_from_array(
  file,
  data,
  var_names = NULL,
  has_T_timeAxis,
  has_Z_verticalAxis,
  isGridded = TRUE,
  grid = NULL,
  locations,
  crs,
  force_v4 = TRUE
)
}
\arguments{
\item{file}{A character string. File path to a pre-created, empty netCDF.}

\item{data}{A numeric array. Can be either two or three dimensions. If the data
is two dimensions, The number of columns in the array represent either 1) the
number of variables, 2) values of one variable at multiple time
intervals, or 3) values of one variable at multiple vertical (depth)
intervals. If the data contains values organized by time and depth, depth
should be the third demsion in the array.}

\item{has_T_timeAxis}{A logical value. Indicates that the netCDF has a time
dimension, and that the number of columns in the \var{data} represent time
intervals.}

\item{has_Z_verticalAxis}{A logical value. Indicates that the netCDF has a Z
dimension, and that either, if \var{has_T_timeAxis} is \code{FALSE} the
number of columns in the \var{data} represent vertical intervals, or
if \var{has_T_timeAxis} is \code{TRUE} the length of the third dimension in
\var{data} is equal to the length of the vertical axis.}

\item{grid}{filename (character). File containing the grid information
(i.e. resolution, extent) of the data. Supported file types are the 'native'
raster package format and those that can be read via rgdal.}

\item{locations}{A SpatialPoints object or a matrix or data.frame with two
columns containing long and lat values. Data must be organized by long, lat.}

\item{force_v4}{A logical value. Force version 4 of netCDF.}

\item{var_name}{A character string equal to the length of the number of variables.
The length of var_name should either be 1 if there is either a time and/or
depth dimension \emph{or} equal to the number of columns in the array if
there is no time or depth dimension.}

\item{isGridded.}{A logical value. Represents whether the location data is on
a regualr grid or not.}

\item{crs.}{character or object of class 'CRS'.}
}
\value{
This function is used for the side-effect of filling a file.
}
\description{
This function add values to a strucutured netCDF. The empty netCDF would have
been created and strucutred with the data in mind.
}
\examples{

}
