% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_functions_miscellaneous.R
\name{prepare_colors}
\alias{prepare_colors}
\title{Prepare colors for plots}
\usage{
prepare_colors(type, zlim, col_rev = FALSE, col_zero = "gray")
}
\description{
Prepare colors for plots
}
