% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{read_netCDF_to_raster}
\alias{read_netCDF_to_raster}
\title{Enhance \code{\link[raster]{raster}} to handle added information written
to a \var{netCDF} file by function
\code{\link{create_netCDF_from_raster_with_variables}}}
\usage{
read_netCDF_to_raster(x, ...)
}
\value{
A Raster* object.
}
\description{
Enhance \code{\link[raster]{raster}} to handle added information written
to a \var{netCDF} file by function
\code{\link{create_netCDF_from_raster_with_variables}}
}
