% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_change.R
\name{remap_directions_6to2}
\alias{remap_directions_6to2}
\title{Remap from six categories of directional modification to two categories}
\usage{
remap_directions_6to2(data, expected_dy_lt_dx = TRUE, cl = NULL)
}
\arguments{
\item{data}{Integer, factor, or character vector coded according to
\code{define_direction_2deltas()[, "name"]}}
}
\value{
An object like \code{data} but coded according to
  \code{define_direction_2deltas()[, "expected"]}
}
\description{
Remap from six categories of directional modification to two categories
}
