% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2analysis-defunct.R
\name{rSW2analysis-defunct}
\alias{rSW2analysis-defunct}
\alias{calculate_cell_area}
\alias{calculate_nominal_resolution}
\alias{create_raster_from_variables}
\alias{get_isoline_polygon}
\alias{variogram_range}
\alias{create_empty_netCDF_file}
\alias{populate_netcdf_from_array}
\alias{read_netCDF_to_array}
\alias{read_netCDF_to_raster}
\title{Defunct functions in package \pkg{rSW2analysis}}
\usage{
calculate_cell_area(...)

calculate_nominal_resolution(...)

create_raster_from_variables(...)

get_isoline_polygon(...)

variogram_range(...)

create_empty_netCDF_file(...)

populate_netcdf_from_array(...)

read_netCDF_to_array(...)

read_netCDF_to_raster(...)
}
\description{
Executing a defunct function will fail and tell you which function
replaces them.
}
