% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2analysis-package.R
\docType{package}
\name{rSW2analysis}
\alias{rSW2analysis}
\alias{rSW2analysis-package}
\title{\pkg{rSW2analysis}: Tools to extract and analyze output from simulation
 experiments that were carried out by \pkg{rSFSW2}}
\description{
Tools to support analysis and access of outputs of
  a simulation experiment that was executed by rSFSW2 or rSFSTEP2.
}
\section{LICENSE}{

   Copyright (C) \Sexpr{format(Sys.Date(), "\%Y")} by
   \Sexpr{packageDescription("rSFSW2")[["Maintainer"]]}

   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, version 3 of the License.
}

\section{DISCLAIMER}{

   This program is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.
}

\author{
\strong{Maintainer}: Daniel Schlaepfer \email{daniel.schlaepfer@yale.edu} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})

}
