% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geographical.R
\name{SREX2012_regions}
\alias{SREX2012_regions}
\title{Define (adjusted) analysis regions based on IPCC 2012 SREX
Figure 3-1 and Table 3.A-1}
\usage{
SREX2012_regions(adjusted = FALSE)
}
\description{
Define (adjusted) analysis regions based on IPCC 2012 SREX
Figure 3-1 and Table 3.A-1
}
\references{
IPCC (2012) Managing the Risks of Extreme Events and Disasters to Advance
Climate Change Adaptation. A Special Report of Working Groups I and II of the
Intergovernmental Panel on Climate Change  [C. B. Field, V. Baros, T. F.
Stocker, D. Qin, D. J. Dokken, K. L. Ebi, M. D. Mastrandrea, K .J. Mach,
G.-K. Plattner, S. K. Allen, M. Tignor and P. M. Midgley (eds.)], Cambridge
University Press, Cambridge, United Kingdom, and New York, NY, USA.
}
