% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_mathematical.R
\name{whittaker_additive_elements}
\alias{whittaker_additive_elements}
\title{Whittaker, J. 1984. Model Interpretation from the Additive Elements of the Likelihood Function. Journal of the Royal Statistical Society. Series C (Applied Statistics) 33:52-64.
Residual: G(:1234) = g_1234 = residual variation
Primaries: G(1:234) = g1_234 = variation uniquely attributable to x1
Secondaries: G(12:34) = g12_34 = variation attributable to x1 or x2 but not to both
3rd order elements: G(123:4) = g123_4 =
4th order elements: G(1234:) = g1234_ =}
\usage{
whittaker_additive_elements(
  data,
  y = "y",
  w = NULL,
  X1 = "Region",
  X2 = "Shift",
  X3 = "GCM",
  X4 = "RCP",
  family = stats::gaussian()
)
}
\description{
Whittaker, J. 1984. Model Interpretation from the Additive Elements of the Likelihood Function. Journal of the Royal Statistical Society. Series C (Applied Statistics) 33:52-64.
Residual: G(:1234) = g_1234 = residual variation
Primaries: G(1:234) = g1_234 = variation uniquely attributable to x1
Secondaries: G(12:34) = g12_34 = variation attributable to x1 or x2 but not to both
3rd order elements: G(123:4) = g123_4 =
4th order elements: G(1234:) = g1234_ =
}
\examples{
## Test code to match up with cement example of Whittaker 1984
if (requireNamespace("MuMIn")) {
  probs_cement <- whittaker_additive_elements(
    data = MuMIn::Cement,
    X1 = "X1", X2 = "X2", X3 = "X3", X4 = "X4"
  )$probs
}

#            Residuals           Primary_X1           Primary_X2
#                  1.8                  1.0                  0.1
#           Primary_X3           Primary_X4      Secondary_X1:X2
#                  0.0                  0.0                  3.6
#      Secondary_X1:X3      Secondary_X1:X4      Secondary_X2:X3
#                 29.3                 12.6                  0.9
#      Secondary_X2:X4      Secondary_X3:X4    Tertiary_X1:X2:X3
#                 43.3                  0.4                 -4.1
#    Tertiary_X1:X2:X4    Tertiary_X1:X3:X4    Tertiary_X2:X3:X4
#                  9.1                -11.6                  0.2
# Quartery_X1:X2:X3:X4
#                 13.5

}
