% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geographical.R
\name{calculate_cell_area}
\alias{calculate_cell_area}
\title{Calculate areal extent of cells for each site}
\usage{
calculate_cell_area(sp_sites, sim_raster, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{sp_sites}{A \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}}
object for which areal extent of cells will be computed.}

\item{sim_raster}{A \code{\link[raster:Raster-class]{raster::Raster}}
object used for coordinate system and cells. The coordinate system must
use angular coordinates (longitude/latitude).}

\item{tol}{A numerical value to inform the site to cell matching.}
}
\value{
A \code{data.frame} with four columns: \var{\dQuote{Longitude}},
  \var{\dQuote{Latitude}}, \var{\dQuote{km2}}, and \var{\dQuote{rel}};
  and one row for each site of \code{sp_sites}.
  Cell area on ellipsoid, based on \code{sim_raster}, in square-kilometer
  \code{km2} and as fraction of maximal cell area on the equator \code{rel}.
}
\description{
Calculate areal extent of cells for each site
}
\seealso{
\code{\link[geosphere]{areaPolygon}}
}
