% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_raster.R
\name{create_raster_from_variables}
\alias{create_raster_from_variables}
\title{Converts values into a Raster*}
\usage{
create_raster_from_variables(SFSW2_prj_meta, data, filename = "")
}
\arguments{
\item{SFSW2_prj_meta}{An environment}

\item{data}{A vector or two-dimensional object.}

\item{filename}{A character string. Passed to \code{\link[raster]{brick}}.}
}
\value{
A RasterLayer (if \code{data} is a vector) or RasterBrick (if
  \code{data} is two-dimensional).
}
\description{
Converts values into a Raster*
}
