% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2analysis-defunct.R
\name{rSW2analysis-defunct}
\alias{rSW2analysis-defunct}
\alias{calculate_cell_area}
\alias{calculate_nominal_resolution}
\title{Defunct functions in package \pkg{rSW2analysis}}
\usage{
calculate_cell_area(...)

calculate_nominal_resolution(...)
}
\description{
Executing a defunct function will fail and tell you which function
replaces them.
}
